+++ src/prefabs/BombPrefab.ts
@@ -7,39 +7,41 @@ import Phaser from "phaser";
 /* START-USER-IMPORTS */
 /* END-USER-IMPORTS */
 
-export default interface eggPrefab {
+export default interface BombPrefab {
 
 	 body: Phaser.Physics.Arcade.Body;
 }
 
-export default class eggPrefab extends Phaser.GameObjects.Image {
+export default class BombPrefab extends Phaser.GameObjects.Image {
 
 	constructor(scene: Phaser.Scene, x?: number, y?: number, texture?: string, frame?: number | string) {
 		super(scene, x ?? 0, y ?? 0, texture || "bird1egg", frame);
 
 		scene.physics.add.existing(this, false);
-		this.body.gravity.y = 100;
 		this.body.setSize(11, 12, false);
 
 		/* START-USER-CTR-CODE */
-
+		// Write your code here.
 		/* END-USER-CTR-CODE */
 	}
 
 	/* START-USER-CODE */
 
-	enable()
+	/** asdf */
+	appear()
 	{
 		this.setActive(true);
 		this.setVisible(true);
 		this.body.setEnable(true);
+		// this.body.reset(this.x, this.y);
 	}
 
-	disable()
+	/** asdf */
+	disappear()
 	{
 		this.setActive(false);
-		this.body.setEnable(false);
 		this.setVisible(false);
+		this.body.setEnable(false);
 	}
 
 	/* END-USER-CODE */